+++ Yedam_Team_Two_Project/src/main/webapp/static/css/privateGoods.css
@@ -1,0 +1,109 @@
+/* .single-products-catagory1 요소 */
+.single-products-catagory1 {
+    position: relative;
+    z-index: 1;
+    width: auto; /* 이미지 크기에 제약을 가하지 않음 */
+    top: 0;
+}
+
+@media only screen and (min-width: 992px) and (max-width: 1199px) {
+    .single-products-catagory1 {
+        width: 50%;
+    }
+}
+
+@media only screen and (min-width: 768px) and (max-width: 991px) {
+    .single-products-catagory1 {
+        width: 100%;
+    }
+}
+
+@media only screen and (max-width: 767px) {
+    .single-products-catagory1 {
+        width: 100%;
+    }
+}
+
+@media only screen and (min-width: 576px) and (max-width: 767px) {
+    .single-products-catagory1 {
+        width: 50%;
+    }
+}
+
+.single-products-catagory1 a {
+    display: block;
+    width: 100%;
+    position: relative;
+    z-index: 10;
+}
+
+.single-products-catagory1 a::after {
+    -webkit-transition-duration: 250ms;
+    transition-duration: 250ms;
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    top: 0;
+    left: 0;
+    content: '';
+    background-color: rgba(60, 60, 60, 0.7);
+    z-index: 1;
+    opacity: 0;
+    visibility: hidden; /* 기본적으로는 숨김 */
+}
+
+.single-products-catagory1:hover a::after,
+.single-products-catagory1:focus a::after {
+    opacity: 1;
+    visibility: visible; /* 호버할 때 보이도록 변경 */
+}
+
+.single-products-catagory1 a img {
+    width: 100%;
+    height: auto; /* 이미지의 높이를 자동으로 조정 */
+}
+
+.single-products-catagory1 .hover-content1 {
+    position: absolute;
+    top: 0;
+    left: 0;
+    z-index: 10;
+}
+
+@media only screen and (max-width: 767px) {
+    .single-products-catagory1 .hover-content1 {
+        top: 0;
+        left: 0;
+    }
+}
+
+.single-products-catagory1 .hover-content1 .line {
+    width: 80px;
+    height: 3px;
+    background-color: #fbb710;
+    display: block;
+    margin-bottom: 15px;
+}
+
+.single-products-catagory1 .hover-content1 p {
+    font-size: 16px;
+    color: #6d6d6d;
+    line-height: 1;
+    margin-bottom: 5px;
+    transition-duration: 250ms;
+    visibility: hidden; /* 기본적으로는 숨김 */
+}
+
+.single-products-catagory1 .hover-content1 h4 {
+    margin-bottom: 0;
+    transition-duration: 250ms;
+    visibility: hidden; /* 기본적으로는 숨김 */
+}
+
+.single-products-catagory1:hover .hover-content1 p,
+.single-products-catagory1:hover .hover-content1 h4,
+.single-products-catagory1:focus .hover-content1 p,
+.single-products-catagory1:focus .hover-content1 h4 {
+    color: #fff;
+    visibility: visible; /* 호버할 때 보이도록 변경 */
+}
